Coming back after over a year
The generated parser does not work and I am no longer altogether accustomed to the project structure, so I'm saving the current state of the code and will attempt to amend a few things.  Previously I had been working on the syntax tree transformer to convert the parse tree for an Umbrella source file into a collection of data structures suitable for interpretation.  That is about halfway done but before finishing it I need to fix the generated parser and so I will be rewriting the tup file so that I can get coverage data and write unit tests.